# packages/server/docs/swagger.yaml
basePath: /
definitions:
  github_com_mhmadamrii_kommers_server_internal_model.DiscountType:
    enum:
    - percent
    - fixed
    type: string
    x-enum-varnames:
    - DiscountTypePercent
    - DiscountTypeFixed
...
  internal_handler.addCartItemRequest:
    properties:
      product_id:
        type: integer
      quantity:
        type: integer
    required:
    - product_id
    - quantity
    type: object
...
  internal_handler.campaignRequest:
    properties:
      category_id:
        description: |-
          Exactly one of these two must be set — a campaign targets either a
          whole category or an explicit list of products, never both.
        type: integer
      discount_type:
        allOf:
        - $ref: '#/definitions/github_com_mhmadamrii_kommers_server_internal_model.DiscountType'
        enum:
        - percent
        - fixed
      discount_value:
        type: integer
      ends_at:
        type: string
      free_shipping:
        type: boolean
      name:
        type: string
      product_ids:
        items:
          type: integer
        type: array
      starts_at:
        type: string
      stock_limit:
        type: integer
    required:
    - discount_type
    - discount_value
    - ends_at
    - name
    - starts_at
    type: object
  internal_handler.campaignResponse:
    properties:
      category_id:
        type: integer
      created_at:
        type: string
      discount_type:
        $ref: '#/definitions/github_com_mhmadamrii_kommers_server_internal_model.DiscountType'
      discount_value:
        type: integer
      ends_at:
        type: string
      free_shipping:
        type: boolean
      id:
        type: integer
      is_active:
        type: boolean
      name:
        type: string
      product_ids:
        items:
          type: integer
        type: array
      starts_at:
        type: string
      stock_limit:
        type: integer
      stock_used:
        type: integer
    type: object
  internal_handler.cartItemResponse:
    properties:
      id:
        type: integer
      image_url:
        type: string
      price_cents:
        type: integer
      product_id:
        type: integer
      product_name:
        type: string
      product_slug:
        type: string
      quantity:
        type: integer
      subtotal_cents:
        type: integer
    type: object
  internal_handler.cartResponse:
    properties:
      id:
        type: integer
      items:
        items:
          $ref: '#/definitions/internal_handler.cartItemResponse'
        type: array
      total_cents:
        type: integer
    type: object
...
  internal_handler.updateCartItemRequest:
    properties:
      quantity:
        type: integer
    required:
    - quantity
    type: object
...
info:
  contact: {}
  description: E-commerce backend API for kommers.
  title: Kommers API
  version: "1.0"
paths:
...
  /api/v1/campaigns:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/internal_handler.campaignResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List campaigns (admin sees all, seller sees their own)
      tags:
      - campaigns
    post:
      consumes:
      - application/json
      parameters:
      - description: Campaign payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_handler.campaignRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal_handler.campaignResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 'Create a promotion campaign (admin: any target; seller: own products
        only)'
      tags:
      - campaigns
  /api/v1/campaigns/{id}:
    delete:
      parameters:
      - description: Campaign ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Cancel a campaign (admin, or the owning seller)
      tags:
      - campaigns
  /api/v1/cart:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler.cartResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get the current user's cart
      tags:
      - cart
  /api/v1/cart/items:
    post:
      consumes:
      - application/json
      parameters:
      - description: Item payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_handler.addCartItemRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler.cartResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Add a product to the cart (or increase quantity if already present)
      tags:
      - cart
  /api/v1/cart/items/{id}:
    delete:
      parameters:
      - description: Cart item ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Remove an item from the cart
      tags:
      - cart
    put:
      consumes:
      - application/json
      parameters:
      - description: Cart item ID
        in: path
        name: id
        required: true
        type: integer
      - description: Quantity payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_handler.updateCartItemRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler.cartResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update a cart item's quantity
      tags:
      - cart
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and the JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"